// src/models/Project.ts
export type Project = {
    name: string;
    description: string;
    type: string;
    amount: number;
    amountRaised: number;
    dateEnd: Date;
    dateStart: Date;
    id: string;
    image: string;
    interestRate: number;
    location: string;
    owner: string;
    paymentMethod: string; 
    paymentType: string;
    risk: string;
    timeLeft: number;
    totalInvestors: number;
    video: string;
}